requirements.txt
pygame-ce
Pillow
pywin32
numpy